Setting up NextJS API
